+++ src/main/java/Model/Entitys/Monsters/Bird.java
@@ -8,7 +8,7 @@ public class Bird extends AbstractMonster {
 
     private boolean totop = true;
 
-    public Bird(Etage m, Position pos, String nom, int pv, int force, int vision_radius , int agro, int update_rate, int pathCross) {
+    public Bird(Etage m, Position pos, String nom, int pv, int force, double vision_radius , int agro, int update_rate, int pathCross) {
         super(m, pos, nom, pv, force, vision_radius, agro, update_rate, pathCross);
     }
