+++ src/main/java/Model/Entitys/Monsters/Zombie.java
@@ -5,7 +5,7 @@ import Model.Utils.Affichage;
 import Model.Utils.Position;
 
 public class Zombie extends AbstractMonster {
-    public Zombie(Etage m, Position pos, String nom, int pv, int force, int vision_radius , int agro, int update_rate, int pathCross) {
+    public Zombie(Etage m, Position pos, String nom, int pv, int force, double vision_radius , int agro, int update_rate, int pathCross) {
         super(m, pos, nom, pv, force, vision_radius, agro, update_rate, pathCross);
     }
